gmscompat: don't use "/proc/self/fd" paths for loading Dynamite modules
dup()-ing fd that is referred to by "/proc/self/fd" path isn't the same
as open()-ing that path. fd that is returned by dup() refers to the same
underlying file description and thus shares position and file status
flags with the original fd.

The ability to pass "/proc/self/fd" path instead of a file path
is an undocumented API on Android, but there are apps that depend on it.

Signed-off-by: Escobar1945 <candaroglu197878@gmail.com>